WIP: force tab web process initialization
This commit attempts to fix the broken UI issues starting with
webkit2gtk 2.26. These issues are caused by changes in the worker
process model.

This commit now forces a worker process to be started when a tab is
switched to by loading an empty string of HTML. Subsequent loads are
then blocked using the existing load block mechanism, and are unblocked
once the web extensions have been loaded.

This produces an extra entry in the tab history. Some effort has been
made to filter such entries out in C, but I suspect it's currently
insufficient. This loading state is also visible through the view URI;
this would be avoidable by moving this initialization logic entirely to
C.